# docs/swagger.yaml
basePath: /api
definitions:
  models.Client:
    properties:
      balance:
        minimum: 0
        type: integer
      id:
        type: string
    required:
    - id
    type: object
  models.Transaction:
    properties:
      balance:
        minimum: 0
        type: integer
      client_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      status:
        maxLength: 255
        type: string
      transaction_type:
        maxLength: 255
        type: string
    required:
    - client_id
    - id
    - status
    - transaction_type
    type: object
info:
  contact:
    email: your@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API
  version: "1.0"
paths:
...
  /v1/client/{id}:
    get:
      consumes:
      - application/json
      description: Get client by given ID.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
      summary: get client by given ID
      tags:
      - Client
...
  /v1/transaction/{id}:
    get:
      consumes:
      - application/json
      description: Get transaction by given ID.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      summary: get transaction by given ID
      tags:
      - Transaction
...
swagger: "2.0"